A new approach for bootstrapping the register machine.
We use a 5 state Turing maching to leave many alternating 1s and 0s on the tape,
then use a small second stage bootstrap to repeatedly increment and decrement
register 0, the decrement adding an additional register to the file each time.

The expanding register file makes the machine even more horribly inefficient,
but it's the state count that matters.