Base zf2.py on the zf2.nql version by sorear.
Update zf2.py to use the axioms and axiom order from:
<https://github.com/sorear/metamath-turing-machines/blob/master/zf2.nql>

GEN and DET cannot appear last in the axiomcode assignments. They can
produce bogus theorems for axiomcodes greater than their assigned axiomcode
because cparam() won't fire, and it is critical for there to be a subsequent
axiom to discard the bad theorem from wfftop.